Reverts addon/app file position
